remove `parser/common.rs`
